made expiriments and fixed bugs

+++ DoubleHashTable.java
@@ -13,7 +13,13 @@ public class DoubleHashTable extends OAHashTable {
 
 	@Override
 	public int Hash(long x, int i) {
-		return (this.h1.Hash(x) + i * (this.h2.Hash(x) + 1)) % m;
+		// long a = this.h1.Hash(x);
+		// long b = this.h2.Hash(x) + 1;
+		// long j = i;
+		// long val = (a + (j * b)) % m;
+		// int index = (int) val;
+		// return index;
+		return (int) (((long) this.h1.Hash(x) + ((long) i * (long) (this.h2.Hash(x) + 1))) % m);
 	}
 
 }
